Not finished -- polymorphism to pointer

+++ DM2231_Lab2_MVC/DM2231_Lab2/Mouse.h
@@ -14,7 +14,7 @@ public:
 	bool bLButtonUp,bRButtonUp;
 
 	
-	Vector3D<int> MousePos;
+	Vector3D<float> MousePos;
 
 	void Init(void)
 	{
